refactor create utilities and move saveRecors to firebaseConfig.js
